Reduce memory usage
It would read the entire storage tail, and then copy it into a numpy
object, and then into a pandas dataframe, and then it would clip it to
max_records. Now it reads the entire storage tail, then clips it and
then does the rest. I had a storage tail of 1.8 million and
max_records 100 thousand, and this improvement reduced memory usage from
more than 8G to 1G. It also doesn't grow too much when it has several
configuration sections to process, whereas before, apparently because of
memory fragmentation, the memory for each station was adding up to the
one already used.